♻️ Move Recommendation Skeleton to Recommendation component
